fixed a bug related to laval buckets and similar fuel
When you have a fuel enderchest with lava buckets, the turtles may sometime get an empty bucket when accessing it and get stuck because they do not put the empty bucket back before checking for new fuel. Simply ordering it to dropUp and then wait a second before sucking in new fuel from the chest should solve this problem.